+++ demo2/src/components/modal/board/BoardDetail.vue
@@ -1,0 +1,153 @@
+<template>
+  <div
+    class="modal fade"
+    id="boardDetailModal"
+    tabindex="-1"
+    aria-hidden="true"
+  >
+    <div
+      class="modal-dialog modal-xl modal-danger modal-dialog-centered modal-"
+      role="document"
+    >
+      <div class="modal-content">
+        <div class="modal-body p-0">
+          <div class="modal-content">
+            <div class="modal-body">
+                <section class="pt-3 pt-md-5 pb-md-5 mt-8 mt-md-7 mt-lg-0">
+                <div class="container">
+                    <div class="col-12 mx-auto pb-5">
+                    <div class="card shadow-lg">
+                        <div
+                        class="
+                            card-header
+                            p-0
+                            position-relative
+                            mt-n4
+                            mx-3
+                            z-index-2
+                            bg-transparent
+                        "
+                        >
+                        <div
+                            class="
+                            bg-gradient-dark
+                            shadow-dark
+                            border-radius-lg
+                            p-4
+                            "
+                        >
+                            <h3 class="text-white mb-0">제목</h3>
+                            <p class="text-white opacity-8 mb-0">
+                            작성 일자: Sept 01 2021
+                            </p>
+                        </div>
+                        </div>
+                        <div class="card-body p-5 pt-0">
+                        <h5 class="mt-5 mb-3">Introduction</h5>
+                        At Creative Tim, accessible
+                        <a href="https://www.creative-tim.com/">here</a>, one of
+                        our main priorities is the privacy of our visitors. This
+                        Privacy Policy document contains types of information that
+                        is collected and recorded by Website Name and how we use
+                        it. <br /><br />
+                        If you have additional questions or require more
+                        information about our Privacy Policy, do not hesitate to
+                        contact us through email at
+                        <a
+                            href="/cdn-cgi/l/email-protection"
+                            class="__cf_email__"
+                            data-cfemail="5e363b3232311e3d2c3b3f2a37283b732a3733703d3133"
+                            >[email&#160;protected]</a
+                        >
+                        <br /><br />
+                        This privacy policy applies only to our online activities
+                        and is valid for visitors to our website with regards to
+                        the information that they shared and/or collect in
+                        <a href="https://www.creative-tim.com/">Creative Tim</a>.
+                        This policy is not applicable to any information collected
+                        offline or via channels other than this website.
+                        <h5 class="mt-5 mb-3">Managing Your Information</h5>
+                        <p>
+                            Unless otherwise stated, Creative Tim and/or its
+                            licensors own the intellectual property rights for all
+                            material on Creative Tim. All intellectual property
+                            rights are reserved. You may access this from Creative
+                            Tim for your own personal use subjected to restrictions
+                            set in these terms and conditions.
+                        </p>
+                        <p>You must not:</p>
+                        <ul>
+                            <li>Republish material from Creative Tim</li>
+                            <li>
+                            Sell, rent or sub-license material from Creative Tim
+                            </li>
+                            <li>
+                            Reproduce, duplicate or copy material from Creative
+                            Tim
+                            </li>
+                            <li>Redistribute content from Creative Tim</li>
+                        </ul>
+                        <p>
+                            This Agreement shall begin on the date hereof. Our Terms
+                            and Conditions were created with the help of the
+                            <a href="https://www.creative-tim.com"
+                            >Terms And Conditions Generator</a
+                            >
+                            and the
+                            <a href="https://www.creative-tim.com"
+                            >Privacy Policy Generator</a
+                            >.
+                        </p>
+                        <p>
+                            Parts of this website offer an opportunity for users to
+                            post and exchange opinions and information in certain
+                            areas of the website. Creative Tim does not filter,
+                            edit, publish or review Comments prior to their presence
+                            on the website. Comments do not reflect the views and
+                            opinions of Creative Tim,its agents and/or affiliates.
+                            Comments reflect the views and opinions of the person
+                            who post their views and opinions. To the extent
+                            permitted by applicable laws, Creative Tim shall not be
+                            liable for the Comments or for any liability, damages or
+                            expenses caused and/or suffered as a result of any use
+                            of and/or posting of and/or appearance of the Comments
+                            on this website.
+                        </p>
+                        <p>
+                            Creative Tim reserves the right to monitor all Comments
+                            and to remove any Comments which can be considered
+                            inappropriate, offensive or causes breach of these Terms
+                            and Conditions.
+                        </p>
+                        <h5 class="mt-5 mb-3">Security</h5>
+                        We shall not be hold responsible for any content that
+                        appears on your Website. You agree to protect and defend
+                        us against all claims that is rising on your Website. No
+                        link(s) should appear on any Website that may be
+                        interpreted as libelous, obscene or criminal, or which
+                        infringes, otherwise violates, or advocates the
+                        infringement or other violation of, any third party
+                        rights.
+                        </div>
+                    </div>
+                    </div>
+                </div>
+                </section>
+            </div>
+            <div class="modal-footer">
+                <button type="button" class="btn bg-gradient-secondary" data-bs-dismiss="modal">닫기</button>
+                <button type="button" class="btn bg-gradient-primary">수정</button>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
+
+<script>
+export default {};
+</script>
+
+<style>
+</style>
